Add the ability to react to message content, or to react to reactions. Each message/reaction trigger configuration now has a new supported key "REACTIONS", which can be defined. This will instruct the bot to choose a reaction from that list each time an associated trigger was found.
Some functionality was abstracted to an async helper function because much of the logic could be shared between the message triggering and reaction triggering.

In order to make the trigger search more generic to messages or reactions, some utils.core code was modified, and all associated uses of the changed core functions have been updated and tested.